requirements.txt
africastalking==1.2.7
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
Authlib==1.3.0
blinker==1.7.0
cachelib==0.12.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contextlib2==21.6.0
coverage==7.4.4
cryptography==42.0.5
flasgger==0.9.7.1
Flask==2.3.3
Flask-Dance
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-OAuthlib
flask-oidc==2.1.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib_resources==6.3.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
mistune==3.0.2
oauthlib
packaging==24.0
psycopg-binary==3.1.18
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.4.0
rpds-py==0.18.0
rsa==4.9
schema==0.7.5
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
URLObject==2.4.3
Werkzeug==3.0.1
